Refactor result
