+++ web/frontend/src/Routes.ts
@@ -7,9 +7,7 @@ export const ROUTE_ABOUT = '/about';
 export const ROUTE_ELECTION_INDEX = '/election/index';
 export const ROUTE_ELECTION_CREATE = '/election/create';
 export const ROUTE_ELECTION_SHOW = '/elections/:electionID';
-
-export const ROUTE_RESULT_INDEX = '/result/index';
-export const ROUTE_RESULT_SHOW = '/result/show';
+export const ROUTE_ELECTION_RESULT = '/elections/:electionID/result';
 
 export const ROUTE_BALLOT_INDEX = '/ballot/index';
 export const ROUTE_BALLOT_CAST = '/ballot/cast';
