+++ web/frontend/src/components/utils/useElection.tsx
@@ -1,15 +1,15 @@
 import useFetchCall from './useFetchCall';
 import * as endpoints from './Endpoints';
 import { useFillElectionInfo } from './FillElectionInfo';
+import { ID } from 'types/configuration';
 
 // Custom hook that fetches an election given its id and returns its
 // different parameters
-const useElection = (electionID) => {
+const useElection = (electionID: ID) => {
   const request = {
     method: 'GET',
   };
-
-  const [data, loading, error] = useFetchCall(endpoints.election(electionID.toString()), request);
+  const [data, loading, error] = useFetchCall(endpoints.election(electionID), request);
   const {
     status,
     setStatus,
