+++ web/frontend/src/pages/election/components/ProgressBar.tsx
